+++ Cargo.toml
@@ -174,7 +174,7 @@ ransac = { path = "crates/ransac" }
 regex = "1.6.0"
 repository = { path = "crates/repository" }
 reqwest = { version = "0.11.23", features = ["blocking"] }
-rmp-serde = "*"
+rmp-serde = "1.3.0"
 rustfft = "6.0.1"
 scenario = { path = "crates/scenario" }
 semver = "1.0.20"
